emacs/bqn-input.el: fix reliance on undefined vars und functions
bqn--make-bqn-mode-map depended on the value of bqn-mode-map-prefix, but
bqn-mode-map-prefix's set function also used bqn--make-bqn-mode-map. To
break this cycle which made emacs' warnings act up a bit, we remove
bqn--make-bqn-mode-map and defvar bqn--mode-map to nil initially,
trusting that it'll be correctly re-defined by the initialization of
bqn-mode-map-prefix via its set function.

I verified that this works correctly both in the case of the initial
value (although I swapped out the value for "M-" beforehand since in my
setup Super doesn't seem to work at all) and redefining it using
costumize-variable.