consolidate inferior-process sending
The typical use case for inferior-process sending is to send many
snippets of code to the inferior process while staying in the source
buffer.  Switching to the inferior process is the exception rather than
the rule; this naturally justifies the optional prefix argument.

This change obviates the `-and-follow' variants, therefore they are
removed.

In the course of these changes, avoid user commands to operate on
buffers, and when switching buffers, resort to commands that respect
user customizations.

Also for a more "do-what-I-mean" behavior, when a region is active, send
that region instead of the current line.

Being escaped, the sending now works with multiple lines.