consolidate inferior-process send commands, prefix arg to follow
This removes the `-and-follow' command variants.  When working in a
source buffer, it is common to send many parts to the inferior process,
staying in the source buffer.  Switching to the inferior process is the
exception from that, and a wonderful use of the optional prefix arg.

Also for a more DWIM behavior, when a region is active, send that
instead of the current line.