do not insert source buffer parts to send them
This is the typical usage, especially important when sending larger
chunks (like a whole buffer).  It's even more important now that the
content uses the ")escape" construction.

Avoid sending initial newline, because otherwise the evaluation of an
empty line will be triggered, resulting in a spurious prompt that
destroys output alignment.