separate bqn--glyph-map
Split the special-glyph modifier bindings from the general keymap of bqn
mode.  It can then be re-used for the comint mode, and customization
becomes easier and more robust.

Use "modifier" instead of the word "prefix" in this context; the latter
already has a different meaning for keymaps.

The bindings are fixed to match the input method, so they are not
intended for customization.  Hence the keymap stays internal, and the
commands for symbol insertion can be inlined, this simplifies the code
and does not add 60+ commands that are never used outside the keymap.

I could find no reason why the defcustom initialization was changed from
its default value, which also calls the `:set' function, so use that.

It is very easy now to give the user the option not to add the modifier
bindings at all --- do that when the option is nil.