remove `bqn-help-symbol-at-point-is-called`
Removing an existing command is a tough call.  The argument is that this
command can only possible serve a very special use case and user base:
It gives less (and not particularly useful) information than
either (automatic) eldoc or the full help buffer.

In the particular case that the primitive's name is really that
interesting to a user, it seems fair to me to have the user add a simple
command for that themselves.